[fix]火気使用申請のステータスが変わらない問題の修正
