+++ user/src/components/Applications/FireEquipment/components/hooks.ts
@@ -4,6 +4,10 @@ import {
   useFireEquipmentMutations,
   useGetFireEquipmentOrderByGroupId,
 } from '@/api/fireEquipmentApi';
+import {
+  useGetHealthCenterSubmissionStatus,
+  useUpdateHealthCenterSubmissionStatus,
+} from '@/api/healthCenterSubmissionStatusApi';
 import { NO_ID_STRING, YES_ID_STRING } from '@/utils/constant';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
@@ -18,7 +22,8 @@ import {
 export const useFireEquipmentOrder = (
   groupId: number,
   fireEquipmentData?: FireEquipmentResponse,
-  handleEditCancel?: () => void
+  handleEditCancel?: () => void,
+  status?: string
 ) => {
   const { mutateFireEquipmentOrder } =
     useGetFireEquipmentOrderByGroupId(groupId);
@@ -131,6 +136,29 @@ export const useFireEquipmentOrder = (
     );
   };
 
+  //ステータス変更処理
+  const { healthCenterSubmissionStatus, mutateHealthCenterSubmissionStatus } =
+    useGetHealthCenterSubmissionStatus(groupId);
+  const { trigger: patchHealthCenterSubmissionStatus } =
+    useUpdateHealthCenterSubmissionStatus()();
+
+  const updateStatus = async (status: 'unapproved') => {
+    const fireEquipmentSubmission = healthCenterSubmissionStatus.find(
+      (submission) => submission.applicationType === 'equipment'
+    );
+
+    if (!fireEquipmentSubmission?.id) {
+      throw new Error('Fire equipment submission status id not found');
+    }
+
+    await patchHealthCenterSubmissionStatus({
+      id: fireEquipmentSubmission.id,
+      body: { status },
+    });
+
+    await mutateHealthCenterSubmissionStatus();
+  };
+
   // 火気申請の登録・更新
   const onSubmitFireEquipment = async (formData: FireEquipmentFormValues) => {
     const payload = {
@@ -152,6 +180,13 @@ export const useFireEquipmentOrder = (
         toast.success('火気申請が完了しました！');
       }
       await mutateFireEquipmentOrder();
+
+      // 再提出完了時
+      if (status === 'waiting_resubmission') {
+        // status更新処理
+        await updateStatus('unapproved');
+      }
+
       // 成功後に編集モードを終了
       handleEditCancel?.();
     } catch (error) {
